% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armabottom_k}
\alias{armabottom_k}
\title{Find t smallest index of a vector (RcppArmadillo-version)}
\usage{
armabottom_k(x, k)
}
\arguments{
\item{x}{A vector.}

\item{k}{A int.}
}
\value{
The index of the t smallest element of the vector.
}
\description{
Find t smallest index of a vector (RcppArmadillo-version)
}
