% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaComputeLoss}
\alias{armaComputeLoss}
\title{Compute loss function for OSS (RcppArmadillo-version)}
\usage{
armaComputeLoss(X, xa, y, ya, tPow)
}
\arguments{
\item{X}{Matrix of the candidate set.}

\item{xa}{Norm of the candidate set.}

\item{y}{A vector. The point which be selected last iteration.}

\item{ya}{Norm of \code{y}.}

\item{tPow}{The power of the loss function.}
}
\value{
Loss of the candidate set.
}
\description{
Compute loss function for OSS (RcppArmadillo-version)
}
