% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaIES}
\alias{armaIES}
\title{IES Core Code Using \code{RcppArmadillo}.}
\usage{
armaIES(X, n, q)
}
\arguments{
\item{X}{A data.frame or matrix consists of explanatory variables.}

\item{n}{Subsample size.}

\item{q}{Hyperparamter of how to divide the axes.}
}
\description{
IES Core Code Using \code{RcppArmadillo}.
}
