% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaIES_compare}
\alias{armaIES_compare}
\title{IES C++-Version for Benchmarking (C++ Core Code)}
\usage{
armaIES_compare(X, n, q)
}
\arguments{
\item{X}{A data.frame or matrix consists of explanatory variables.}

\item{n}{Subsample size.}

\item{q}{Hyperparamter of how to divide the axes.}
}
\description{
There is no randomness, all parts that need to be randomly selected are selected first indexed.
}
