% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armarcppIBOSS}
\alias{armarcppIBOSS}
\title{IBOSS with \code{RcppArmadillo} by \verb{the package itself} (\code{myArma_IBOSS} core c++ code)}
\usage{
armarcppIBOSS(X, n)
}
\arguments{
\item{X}{A data.frame or matrix consists of explanatory variables.}

\item{n}{Subsample size.}
}
\value{
Subsample index.
}
\description{
IBOSS with \code{RcppArmadillo} by \verb{the package itself} (\code{myArma_IBOSS} core c++ code)
}
