% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaScaleMatrix}
\alias{armaScaleMatrix}
\title{Scale a matrix (RcppArmadillo-version)}
\usage{
armaScaleMatrix(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
Scaled matrix.
}
\description{
Scale a matrix (RcppArmadillo-version)
}
