% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bottom_t_index}
\alias{bottom_t_index}
\title{Find t smallest index of a vector}
\usage{
bottom_t_index(loss, t)
}
\arguments{
\item{loss}{A vector.}

\item{t}{A int.}
}
\value{
The index of the t smallest element of the vector.
}
\description{
Find t smallest index of a vector
}
