% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IES.R
\name{c_IES_compare}
\alias{c_IES_compare}
\title{IES C++-Version for Benchmarking (R-Wrap Code)}
\usage{
c_IES_compare(X, n, q)
}
\arguments{
\item{X}{X A data.frame or matrix consists of explanatory variables.}

\item{n}{Subsample size.}

\item{q}{Hyperparamter of how to divide the axes.}
}
\description{
There is no randomness, all parts that need to be randomly selected are selected first indexed.
}
