% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeLoss}
\alias{ComputeLoss}
\title{Compute loss function for OSS}
\usage{
ComputeLoss(candi, last_index, X, norm)
}
\arguments{
\item{candi}{The index of the candidate set.}

\item{last_index}{The index of the seleted point in last iteration.}

\item{X}{The whole data.}

\item{norm}{Norm of the whole data.}
}
\value{
Loss of every point in candidate set.
}
\description{
Compute loss function for OSS
}
