% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_IES_Case_2_Train}
\alias{data_IES_Case_2_Train}
\title{Simulation datasets of IES Case2}
\format{
\subsection{data_IES_Case_2_Train}{

A data.frame with 10000 rows and 5 columns.
\describe{
\item{X1-X3}{Explanatory variables generated from truncated multivariate exponential distribution.}
\item{m}{Real overall average.}
\item{y}{Response variable.}
}
}
}
\usage{
data_IES_Case_2_Train
}
\description{
A data.frame with numeric response and real overall mean and three predictors generated from truncated multivariate exponential distribution.
}
\details{
The three predictors generated from truncated multivariate exponential distribution with
covariance matrix \eqn{\Sigma=(0.3^{\mathbb{1}(i \ne j)})}. The marginal distribution is exponential distribution
with rate 1, and is truncated by 4 and translated to \eqn{[-2, 2]}.

The response are generated by \eqn{y = m(x) + \epsilon},  where
\deqn{\begin{aligned} m(x) &= m_1(x_1) + m_2(x_2) + m_3(x_3) \\
                           &= \frac{8}{4 + x_1} + \frac{\exp(3-x^2_2)}{4} + 1.5\sin(\frac{\pi}{2}x_3),
       \end{aligned}}
and \eqn{\epsilon} follows \eqn{N(0, \sigma^2)}, the variance is \eqn{\sigma^2=0.25}.

We generated the above predictors using the \code{ellipCopula} and \code{rCopula} function in package \code{copula}.
}
\keyword{datasets}
