% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsubsampling-package.R
\docType{package}
\name{dbsubsampling-package}
\alias{dbsubsampling}
\alias{dbsubsampling-package}
\title{dbsubsampling: Subsampling Methods Based on Experimental Design}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of design-based subsampling methods which inspired by optimal design, space-filling design, orthogonal array, etc. Provides a unified interface for IBOSS, Lowcon, OSS and other popular methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JieYinStat/dbsubsampling}
  \item \url{https://jieyinstat.github.io/dbsubsampling/}
  \item Report bugs at \url{https://github.com/JieYinStat/dbsubsampling/issues}
}

}
\author{
\strong{Maintainer}: Jie Yin \email{yjabcdhot@hotmail.com}

}
\keyword{internal}
