% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDS.R
\name{DDS}
\alias{DDS}
\title{Data-Driven Subsampling Based on Uniform Design}
\usage{
DDS(X, n, ratio = 0.85)
}
\arguments{
\item{X}{A data.frame or matrix of explanatory variables.}

\item{n}{Subsample size.}

\item{ratio}{Dimensionality reduction ratio of PCA, default to 0.85.}
}
\value{
Subsample index.
}
\description{
A model-free subsampling method based on uniform design proposed by Zhang et.al. (2023).
}
\details{
The uniform design is generated according to the method described in part 5 of Zhang's article.
The uniformity measure uses mixed divergence.
}
\examples{
data <- data_numeric_regression
X <- data[-which(names(data) == "y")]
DDS(X, n = 100, ratio = 0.85)
}
\references{
Mei Zhang, Yongdao Zhou, Zheng Zhou & Aijun Zhang (2023)
\emph{Model-Free Subsampling Method Based on Uniform Designs,
IEEE Transactions on Knowledge and Data Engineering, 36:3, 1210-1220},
\url{https://ieeexplore.ieee.org/abstract/document/10192374}.
}
