% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getIdx_cpp}
\alias{getIdx_cpp}
\title{Get subsample index of the first column (IBOSS core code, Rcpp-C++-style)}
\usage{
getIdx_cpp(r, z)
}
\arguments{
\item{r}{Subsample size of the first column.}

\item{z}{A numeric vector. the first column.}
}
\value{
Subsample index of the first column.
}
\description{
Get subsample index of the first column (IBOSS core code, Rcpp-C++-style)
}
