% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getIdxR_cpp}
\alias{getIdxR_cpp}
\title{Get subsample index of other column(except the first column) (IBOSS core code, Rcpp-C++-style)}
\usage{
getIdxR_cpp(r, z, rdel)
}
\arguments{
\item{r}{Subsample size of the column.}

\item{z}{A numeric vector. the column.}

\item{rdel}{Subsample index of the first column.}
}
\value{
Subsample index of the column.
}
\description{
Get subsample index of other column(except the first column) (IBOSS core code, Rcpp-C++-style)
}
