% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBOSS.R
\name{IBOSS}
\alias{IBOSS}
\title{Information-Based Optimal Subdata Selection for Big Data Linear Regression (IBOSS, Rcpp-c++-style)}
\usage{
IBOSS(n, X)
}
\arguments{
\item{n}{Subsample size.}

\item{X}{A data.frame or matrix consists of explanatory variables.}
}
\value{
Subsample index.
}
\description{
A subsampling method based on D-optiaml criterion inspired by optimal experimental design
used for linear regression.
}
\examples{
data <- data_numeric_regression
X <- data[-which(names(data) == "y")]
IBOSS(n = 100, X = X)
}
\references{
HaiYing Wang, Min Yang & John Stufken (2019)
\emph{Information-Based Optimal Subdata Selection for Big Data Linear Regression,
Journal of the American Statistical Association, 114:525, 393-405},
\url{https://www.tandfonline.com/doi/full/10.1080/01621459.2017.1408468},
\url{https://github.com/Ossifragus/IBOSS}.
}
