% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IES.R
\name{IES}
\alias{IES}
\title{Independence-Encouraging Subsampling for Nonparametric Additive Models (IES, Proposed by Zhang et.al. (2024))}
\usage{
IES(X, n, q = 16, seed = NULL)
}
\arguments{
\item{X}{A data.frame or matrix consists of explanatory variables.}

\item{n}{Subsample size.}

\item{q}{Hyperparamter of how to divide the axes. Default to 16.}

\item{seed}{Random seed for the sampling.}
}
\value{
Subsample index.
}
\description{
A subsampling method for nonparameter additive model based on Orthogonal Array.
}
\examples{
data <- data_numeric_regression
X <- data[-which(names(data) == "y")]
IES(X, n = 100, q = 16, seed = NULL)
}
\references{
Yi Zhang, Lin Wang, Xiaoke Zhang & HaiYing Wang (2024)
\emph{Independence-Encouraging Subsampling for Nonparametric Additive Models,
Journal of Computational and Graphical Statistics},
\url{https://www.tandfonline.com/doi/full/10.1080/10618600.2024.2326136}.
}
