% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{L2norm}
\alias{L2norm}
\title{Get L2 norm}
\usage{
L2norm(X)
}
\arguments{
\item{X}{A matrix or data.frame.}
}
\value{
L2 norm of \code{X}(every row).
}
\description{
Get L2 norm of a matrix or data frame.
}
