% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leverage.R
\name{Leverage}
\alias{Leverage}
\title{Subsampling Based on Leverage Scores}
\usage{
Leverage(X, n, shrinkage_alpha = 1, replace = TRUE, seed = NULL)
}
\arguments{
\item{X}{A data.frame or matrix of explanatory variables.}

\item{n}{Subsample size.}

\item{shrinkage_alpha}{Shrinkage for SLEV, default to 1 (do not shrinkage).}

\item{replace}{With replacement or without replacement, default to TRUE.}

\item{seed}{Random seed for the sampling.}
}
\value{
Subsample index.
}
\description{
A subsampling methods based on leverage scores proposed by Ma et.al. (2015).
}
\examples{
data <- data_numeric_regression
X <- data[-which(names(data) == "y")]
Leverage(X, n = 100, shrinkage_alpha = 0.9, replace = TRUE, seed = NULL)
}
\references{
Ping Ma, Michael W. Mahoney & Bin Yu (2015)
\emph{A Statistical Perspective on Algorithmic Leveraging,
Journal of Machine Learning Research, 16:27, 861−911},
\url{https://jmlr.csail.mit.edu/papers/v16/ma15a.html}.
}
