% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowCon.R
\name{LowCon}
\alias{LowCon}
\title{LowCon: A Design-based Subsampling Approach in a Misspecified Linear Model}
\usage{
LowCon(X, n, theta = 1, space_method = "randomLHS", seed = NULL)
}
\arguments{
\item{X}{A data.frame or matrix of explanatory variables.}

\item{n}{Subsample size.}

\item{theta}{Percentage of data shrinkage. Default to 1.}

\item{space_method}{The generation method of initial space-filling design.}

\item{seed}{Random seed for the sampling.}
}
\value{
Subsample index.
}
\description{
A subsampling method based space-filling design for misspecified linear model proposed by Meng et.al. (2021).
}
\examples{
data <- data_numeric_regression
X <- data[-which(names(data) == "y")]
LowCon(X, n = 100, space_method = "randomLHS", theta = 1, seed = NULL)
}
\references{
Cheng Meng, Rui Xie, Abhyuday Mandal, Xinlian Zhang, Wenxuan Zhong & Ping Ma (2021)
\emph{LowCon: A Design-based Subsampling Approach in a Misspecified Linear Model,
Journal of Computational and Graphical Statistics, 30:3, 694-708},
\url{https://www.tandfonline.com/doi/full/10.1080/10618600.2020.1844215}.
}
