% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSS.R
\name{myArma_OSS}
\alias{myArma_OSS}
\title{OSS (\code{RcppArmadillo}-version by \code{Zhu})}
\usage{
myArma_OSS(n, X)
}
\arguments{
\item{n}{Subsample size.}

\item{X}{A matrix.}
}
\value{
Subsample index.
}
\description{
OSS (\code{RcppArmadillo}-version by \code{Zhu})
}
