% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBOSS.R
\name{myR_IBOSS}
\alias{myR_IBOSS}
\title{IBOSS with \verb{base R} by \verb{the package itself}.}
\usage{
myR_IBOSS(n, X)
}
\arguments{
\item{n}{Subsample size.}

\item{X}{A data.frame or matrix consists of explanatory variables.}
}
\value{
Subsample index.
}
\description{
IBOSS with \verb{base R} by \verb{the package itself}.
}
