% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSS.R
\name{myR_OSS}
\alias{myR_OSS}
\title{OSS with \verb{base R} by \verb{the package itself}.}
\usage{
myR_OSS(n, X)
}
\arguments{
\item{n}{Subsample size.}

\item{X}{A matrix.}
}
\value{
Subsample index.
}
\description{
OSS with \verb{base R} by \verb{the package itself}.
}
