% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBOSS.R
\name{myRcpp_cstyle_IBOSS}
\alias{myRcpp_cstyle_IBOSS}
\title{IBOSS with \code{Rcpp-C++-style} by \verb{the package itself}.}
\usage{
myRcpp_cstyle_IBOSS(n, X)
}
\arguments{
\item{n}{Subsample size.}

\item{X}{A data.frame or matrix consists of explanatory variables.}
}
\value{
Subsample index.
}
\description{
IBOSS with \code{Rcpp-C++-style} by \verb{the package itself}.
}
