% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSMAC.R
\name{OSMAC}
\alias{OSMAC}
\title{Optimal Subsampling for Large Sample Logistic Regression(OSMAC)}
\usage{
OSMAC(X, Y, r1, r2, method = c("mmse", "mvc"), seed_1 = NULL, seed_2 = NULL)
}
\arguments{
\item{X}{A data.frame or matrix of explanatory variables.}

\item{Y}{A numeric vector. Response variable.}

\item{r1}{Sample size for pilot sample.}

\item{r2}{Subsample size.}

\item{method}{Sampling methods:
\itemize{
\item \code{mmse}: A-optimal.
\item \code{mvc}: L-optimal.
}}

\item{seed_1}{Random seed for the first stage sampling.}

\item{seed_2}{Random seed for the second stage sampling.}
}
\value{
A numeric vector with length \code{r2} which represent the subsample index.
}
\description{
A subsampling method based on A- / L- optimal for logistic regression proposed by Wang et.al. (2018).
}
\examples{
data <- data_binary_class
y <- data[["y"]]
x <- data[-which(names(data) == "y")]

OSMAC(X = x, Y = y, r1 = 100, r2 = 5, method="mmse", seed_1 = 123, seed_2 = 456)
OSMAC(X = x, Y = y, r1 = 100, r2 = 5, method="mvc", seed_1 = 123, seed_2 = 456)
}
\references{
HaiYing Wang, Rong Zhu and Ping Ma (2018)
\emph{Optimal Subsampling for Large Sample Logistic Regression, Journal of the American
Statistical Association, 113:522, 829-844},
\url{https://www.tandfonline.com/doi/full/10.1080/01621459.2017.1292914},
\url{https://github.com/Ossifragus/OSMAC}.
}
