% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSS.R
\name{OSS}
\alias{OSS}
\title{Orthogonal subsampling for big data linear regression (OSS, \code{Rcpp}-version by \verb{the package itself})}
\usage{
OSS(n, X)
}
\arguments{
\item{n}{Subsample size.}

\item{X}{A matrix or data frame.}
}
\value{
Subsample index.
}
\description{
A subsampling method based on orthogonal array for linear model.
}
\examples{
data_numeric_regression["y"] <- NULL
X <- as.matrix(data_numeric_regression)
OSS(100, X)

}
\references{
Lin Wang, Jake Elmstedt, Weng Kee Wong & Hongquan Xu (2021)
\emph{Orthogonal subsampling for big data linear regression,
The Annals of Applied Statistics, 15(3), 1273-1290},
\url{https://projecteuclid.org/journals/annals-of-applied-statistics/volume-15/issue-3/Orthogonal-subsampling-for-big-data-linear-regression/10.1214/21-AOAS1462.short?tab=ArticleLink}.
}
