% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IES.R
\name{r_IES}
\alias{r_IES}
\title{R Version of IES for Testing}
\usage{
r_IES(X, n, q, seed = NULL)
}
\arguments{
\item{X}{A data.frame or matrix consists of explanatory variables.}

\item{n}{Subsample size.}

\item{q}{Hyperparamter of how to divide the axes.}

\item{seed}{Random seed for the sampling.}
}
\description{
R Version of IES for Testing
}
