% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSS.R
\name{rbottom_t_index}
\alias{rbottom_t_index}
\title{Find t smallest index of a vector.}
\usage{
rbottom_t_index(loss, t)
}
\arguments{
\item{loss}{A vector.}

\item{t}{A int}
}
\value{
The index of the t smallest element of the vector.
}
\description{
Find t smallest index of a vector.
}
