% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSS.R
\name{rComputeLoss}
\alias{rComputeLoss}
\title{Compute loss function for OSS (r-version)}
\usage{
rComputeLoss(candi, last_index, X, norm, p = ncol(X))
}
\arguments{
\item{candi}{The index of the candidate set.}

\item{last_index}{The index of the seleted point in last iteration.}

\item{X}{The whole data.}

\item{norm}{Norm of the whole data.}

\item{p}{Numbers of columns of the data.}
}
\value{
Loss of every point in candidate set.
}
\description{
Compute loss function for OSS (r-version)
}
