% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_cstyle_IBOSS}
\alias{rcpp_cstyle_IBOSS}
\title{IBOSS with \code{Rcpp-C++-style} by myself (\code{myRcpp_cstyle_IBOSS} core c++ code)}
\usage{
rcpp_cstyle_IBOSS(X, n)
}
\arguments{
\item{X}{A data.frame or matrix consists of explanatory variables.}

\item{n}{Subsample size.}
}
\value{
Subsample index.
}
\description{
IBOSS with \code{Rcpp-C++-style} by myself (\code{myRcpp_cstyle_IBOSS} core c++ code)
}
