% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppIBOSS}
\alias{rcppIBOSS}
\title{IBOSS with \code{Rcpp} (\code{myRcppIBOSS} core c++ code)}
\usage{
rcppIBOSS(X, n)
}
\arguments{
\item{X}{A data.frame or matrix consists of explanatory variables.}

\item{n}{Subsample size.}
}
\value{
Subsample index.
}
\description{
IBOSS with \code{Rcpp} (\code{myRcppIBOSS} core c++ code)
}
