% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppOSS}
\alias{rcppOSS}
\title{OSS \code{Rcpp}-version by \verb{the package itself} (\code{OSS} core code)}
\usage{
rcppOSS(X, n)
}
\arguments{
\item{X}{A matrix.}

\item{n}{Subsample size.}
}
\value{
Subsample index.
}
\description{
OSS \code{Rcpp}-version by \verb{the package itself} (\code{OSS} core code)
}
