% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IES.R
\name{scale01}
\alias{scale01}
\title{Scale data to \eqn{[-1, 1]}}
\usage{
scale01(X)
}
\arguments{
\item{X}{A data.frame or matrix.}
}
\value{
Scaled X.
}
\description{
Scale data to \eqn{[-1, 1]}
}
\examples{
X <- matrix(1:20, 5, 4)
scale01(X)
}
