% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowCon.R
\name{scale_neg_pos_1}
\alias{scale_neg_pos_1}
\title{Transform data to \eqn{[-1,1]^p}}
\usage{
scale_neg_pos_1(X)
}
\arguments{
\item{X}{A data.frame or matrix of explanatory variables.}
}
\value{
Scaled \code{X} to \eqn{[-1,1]^p}.
}
\description{
Transform data to \eqn{[-1,1]^p}
}
\examples{
X <- matrix(rnorm(100), nrow = 20, ncol = 5)
scaled_X <- scale_neg_pos_1(X)
apply(scaled_X, 2, range)
}
