% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unif.R
\name{Unif}
\alias{Unif}
\title{Uniform sampling.}
\usage{
Unif(N, n, seed = NULL, replace = TRUE)
}
\arguments{
\item{N}{Total sample size.}

\item{n}{Subsample size.}

\item{seed}{Random seed which is an integer (default NULL). This random seed is only valid for this sampling and
will not affect the external environment}

\item{replace}{A boolean.
\itemize{
\item \code{TRUE} (the default): sampling with replace.
\item \code{FALSE}: sampling without replace
}}
}
\value{
A numeric vector with length \code{n} which represent the subsample index.
}
\description{
Random sampling with equal probability.
}
\examples{
N <- 1000
n <- 10
Unif(N = 1000, n = 10)
Unif(N = 1000, n = 10)
Unif(N = 1000, n = 10) != Unif(N = 1000, n = 10)

Unif(N = 1000, n = 10, seed = 1)
Unif(N = 1000, n = 10, seed = 123)

Unif(N = 1000, n = 10, seed = 123, replace = TRUE)
Unif(N = 1000, n = 10, seed = 123, replace = FALSE)
}
