% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaOSS}
\alias{armaOSS}
\title{OSS (RcppArmadillo-version)}
\usage{
armaOSS(x, k, tPow = 2)
}
\arguments{
\item{x}{A matrix.}

\item{k}{Subsample size.}

\item{tPow}{The power of the loss function.}
}
\value{
Subsample index.
}
\description{
OSS (RcppArmadillo-version)
}
