% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSMAC.R
\name{get_Logistic_MLE}
\alias{get_Logistic_MLE}
\title{MLE for Logistic regression}
\usage{
get_Logistic_MLE(x, y, w)
}
\arguments{
\item{x}{A matrix of explanatory variables.}

\item{y}{A numeric vector. Response variable.}

\item{w}{A numeric vector. The weight of each sample.}
}
\value{
A list.
\itemize{
\item \code{par} : Parameter estimation.
\item \code{message} : Message during iteration.
\item \code{iter} : Iteration times.
}
}
\description{
Maximum likelihood estimation for logistic regression.
}
