% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling.R
\name{subsampling}
\alias{subsampling}
\title{Get subsample index.}
\usage{
subsampling(
  y_name,
  x_name = NULL,
  data,
  n,
  pilot_n = NULL,
  method = "Unif",
  replace = TRUE,
  seed_1 = NULL,
  seed_2 = NULL,
  na_method = NULL
)
}
\arguments{
\item{y_name}{A character. The name of the response variable in the data frame.}

\item{x_name}{A character. The name of the explanatory variable in the data frame.
Default to all variables except the response variable}

\item{data}{A data frame containing response variables and explanatory variables}

\item{n}{Subsample size.}

\item{pilot_n}{pilot sample size (for some method)}

\item{method}{Subsamling methods:
\itemize{
\item \code{Unif}: Random sampling.
\item \code{OSMAC_A}: A subsampling method based on A-optimal for logistic regression proposed by Wang et.al. (2018).
\item \code{OSMAC_L}: A subsampling method based on L-optimal for logistic regression proposed by Wang et.al. (2018).
}}

\item{replace}{A boolean.
\itemize{
\item \code{TRUE} (the default): sampling with replace.
\item \code{FALSE}: sampling without replace
}}

\item{seed_1}{Random seed for the first stage sampling or Unif.}

\item{seed_2}{Random seed for the second stage sampling.}

\item{na_method}{Method to handle NA.}
}
\value{
A numeric vector with length \code{n} which represent the subsample index.
}
\description{
A unified interface for retrieving subsample indexes
}
\examples{
data <- data_binary_class
subsampling(y_name = "y", data = data, n = 30, method = "Unif", seed_1 = 123)
subsampling(y_name = "y", data = data, n = 30, pilot_n = 100,
  method = "OSMAC_A", seed_1 = 123, seed_2 = 456)
subsampling(y_name = "y", data = data, n = 30, pilot_n = 100,
  method = "OSMAC_L", seed_1 = 123, seed_2 = 456)
}
